+++ steam_pipewire/__init__.py
@@ -1,6 +1,6 @@
 """Steam Audio Isolator - Isolate game audio for clean Steam recording"""
 
-__version__ = '0.3.8'
+__version__ = '0.3.9'
 __author__ = 'crashman79'